pyproject.toml
[project]
name = "ai-web-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
